+++ src/jk_packunpack/Packer1.py
@@ -186,7 +186,7 @@ class Packer1(object):
 					os.chmod(destTarFilePath, chModValueI)	# required as tar will not set the execute bit
 
 			if cmdResult.returnCode != 0:
-				cmdResult.dump(writeFunction=log2.error)
+				cmdResult.dump(printFunc=log2.error)
 				raise Exception("Failed to run 'tar'!")
 
 		return destTarFilePath
@@ -257,7 +257,7 @@ class Packer1(object):
 					os.chmod(destTarFilePath, chModValueI)	# required as tar will not set the execute bit
 
 			if cmdResult.returnCode != 0:
-				cmdResult.dump(writeFunction=log2.error)
+				cmdResult.dump(printFunc=log2.error)
 				raise Exception("Failed to run 'tar'!")
 
 		# ----