+++ src/jk_packunpack/Unpacker.py
@@ -132,7 +132,7 @@ class Unpacker(object):
 			cmdResult = jk_simpleexec.invokeCmd(Unpacker._TAR_PATH, tarArgs, workingDirectory=destDirPath)
 
 			if cmdResult.returnCode != 0:
-				cmdResult.dump(writeFunction=log2.error)
+				cmdResult.dump(printFunc=log2.error)
 				raise Exception("Failed to run 'tar'!")
 	#